+++ packages/auto_updater_macos/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+
+* Add support for Sparkle channels via new `setAllowedChannels` method and the `allowedChannels` delegate.
+
 ## 1.0.0
 
 * First major release.