+++ packages/auto_updater_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+
+* Add support for Sparkle channels via new `setAllowedChannels` method.
+
 ## 1.0.0
 
 * First major release.